Add greeting.js component
