+++ app/controllers/root_controller.rb
@@ -1,4 +1,8 @@
 class RootController < ApplicationController
   def index
   end
+  def hoho
+    @message = Greeting.order("RANDOM()").first.text
+    render json: { message: @message }
+  end
 end
