+++ app/javascript/application.js
@@ -6,6 +6,7 @@ import React from 'react';
 import App from './components/App';
 import { store } from './redux/store';
 import { Provider } from 'react-redux';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
@@ -13,8 +14,11 @@ const root = createRoot(container);
 root.render(
   <React.StrictMode>
     <Provider store={store}>  
-      <App />
-      </Provider>
+      <Router>
+        <Routes>
+          <Route path="/" element={<App />} />
+        </Routes>
+      </Router>
+    </Provider>
   </React.StrictMode>,
 );
-
