+++ app/javascript/components/App.js
@@ -1,22 +1,14 @@
-import { useSelector, useDispatch  } from 'react-redux';
-import React, { useEffect } from 'react';
-import { getmsn} from './gretingSlice';
+import React from 'react';
+import { Greeting } from './greeting';
 
 
 const App = () => {
     
-    const message = useSelector((state) => state.greeting.message);
-    
-    
-    const dispatch  = useDispatch();  
-    useEffect(() => {
-        dispatch(getmsn());
-    }, [dispatch]);
     
   
     return (
         <div>
-            <h1>{message}</h1>
+            <Greeting />
         </div>
     )
 };
