--- app/views/greetings/_form.html.erb
@@ -1,22 +1,0 @@
-<%= form_with(model: greeting) do |form| %>
-  <% if greeting.errors.any? %>
-    <div style="color: red">
-      <h2><%= pluralize(greeting.errors.count, "error") %> prohibited this greeting from being saved:</h2>
-
-      <ul>
-        <% greeting.errors.each do |error| %>
-          <li><%= error.full_message %></li>
-        <% end %>
-      </ul>
-    </div>
-  <% end %>
-
-  <div>
-    <%= form.label :text, style: "display: block" %>
-    <%= form.text_area :text %>
-  </div>
-
-  <div>
-    <%= form.submit %>
-  </div>
-<% end %>
