--- app/views/greetings/_greeting.html.erb
@@ -1,7 +1,0 @@
-<div id="<%= dom_id greeting %>">
-  <p>
-    <strong>Text:</strong>
-    <%= greeting.text %>
-  </p>
-
-</div>
