--- app/views/greetings/_greeting.json.jbuilder
@@ -1,2 +1,0 @@
-json.extract! greeting, :id, :text, :created_at, :updated_at
-json.url greeting_url(greeting, format: :json)
