--- app/views/greetings/edit.html.erb
@@ -1,10 +1,0 @@
-<h1>Editing greeting</h1>
-
-<%= render "form", greeting: @greeting %>
-
-<br>
-
-<div>
-  <%= link_to "Show this greeting", @greeting %> |
-  <%= link_to "Back to greetings", greetings_path %>
-</div>
