--- app/views/greetings/index.html.erb
@@ -1,14 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<h1>Greetings</h1>
-
-<div id="greetings">
-  <% @greetings.each do |greeting| %>
-    <%= render greeting %>
-    <p>
-      <%= link_to "Show this greeting", greeting %>
-    </p>
-  <% end %>
-</div>
-
-<%= link_to "New greeting", new_greeting_path %>
