--- app/views/greetings/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.array! @greetings, partial: "greetings/greeting", as: :greeting
