--- app/views/greetings/new.html.erb
@@ -1,9 +1,0 @@
-<h1>New greeting</h1>
-
-<%= render "form", greeting: @greeting %>
-
-<br>
-
-<div>
-  <%= link_to "Back to greetings", greetings_path %>
-</div>
