--- app/views/greetings/show.html.erb
@@ -1,10 +1,0 @@
-<p style="color: green"><%= notice %></p>
-
-<%= render @greeting %>
-
-<div>
-  <%= link_to "Edit this greeting", edit_greeting_path(@greeting) %> |
-  <%= link_to "Back to greetings", greetings_path %>
-
-  <%= button_to "Destroy this greeting", @greeting, method: :delete %>
-</div>
