--- app/views/greetings/show.json.jbuilder
@@ -1,1 +1,0 @@
-json.partial! "greetings/greeting", greeting: @greeting
