+++ package.json
@@ -8,6 +8,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.2",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1"
   },
